+++ src/app/(root)/task/page.tsx
@@ -1,4 +1,4 @@
-import { TaskEl } from "@/app/components_Task/Task";
+import { TaskList } from "@/app/components_Task/TaskList";
 import { CircleCheck, Star } from "lucide-react";
 import Link from "next/link";
 
@@ -7,14 +7,14 @@ export default function Task() {
     <div className="flex flex-col w-full ">
       <div className="flex flex-row gap-5 mb-3 justify-center">
         <Link href="/task/favorit">
-          <Star />
+          <Star className="fill-amber-300" />
         </Link>
         <Link href="/task/completed">
           <CircleCheck />
         </Link>
       </div>
     <div className="flex w-full flex-col lg:flex-row">
-      <TaskEl />
+      <TaskList/>
       </div>
     </div>
   );
