+++ src/app/api/todoApi.ts
@@ -88,6 +88,21 @@ try{
     }
 )
 
+
+// export const  favoriteTodoApi = createAsyncThunk(
+//     "todoData/ importantTodo",
+//     async(payload:{id:string; isImportant: boolean }, thunkAPI)=>{
+// try{
+//     const response = await axios.patch(`/api/todo/${payload.id}`, {isImportant: payload.isImportant},)
+//     console.log('important', response.data)
+//     return response.data;
+// }catch(error){
+//     const err = error as AxiosError;
+//     return thunkAPI.rejectWithValue(err.response?.data || "Something went wrong" )
+// }
+//     }
+// )
+
 export const deleteTodoApi = createAsyncThunk<string,string>(
     "todoData/deleteTodo",
     async(id:string, thunkAPI)=>{
